Create repainting_living_room_calculator.py
The code first defines the prices of the materials as variables, then it prompts the user for the quantity of each material that user will buy. Then it calculates the total cost of nylon and paint with adding 10% extra paint for wastage and weakener, and adding package cost. It then calculates the total cost for worker by multiplying the hours of work by the hourly rate and adding it to the total cost of materials. Finally it prints the total cost of everything that user needs to pay.